Created test suite for `attribute_factory`
Resolves CR changes for #208
